Can insert Notes into the DB
Inserting notes into the DB is now possible. However, the Title of the note cannot be null because of the layout of the DB. As such all of the LandmarkTitles will just be 'Title'

Will start on congregating the funcvtions together so no space is wased. Easy but I am really tired. ¯\_(ツ)_/¯